Update core-pv version
To run pvws under docker we came across an ipv6 issue that caused the container to crash. This has been mitigated by introducing an environment variable called EPICS_PVA_ENABLE_IPV6. If se to "NO", this then allows pvws container to run correctly. This environment wariable is available in core-pv 4.7.3